requirements.txt
python-telegram-bot
aiohttp
python-dotenv
google-search-results